+++ core/commands/storage/upload/upload/do_pay.go
