+++ core/commands/storage/upload/upload/do_sign_balance.go
