+++ core/commands/storage/upload/upload/do_sign_channel_commit.go
