+++ core/commands/storage/upload/upload/do_sign_guard_contracts.go
